Adding previous work that has been done.
This is an import of the folder where I initially stored the files
related to my work on The Statecraft Blueprint. I asked Claude to do
it's best to organize the files. I haven't really reviewed them because
they are more for historical purposes, I'm focused on the work going
forward.